Add a package export file.
-move source code into src folder
#84vnnm
#84vpdu
