+++ auth_provider/lib/src/AppException.dart
