+++ auth_provider/lib/src/AppleAuthMethod.dart
