+++ auth_provider/lib/src/AuthMethod.dart
