+++ auth_provider/lib/src/AuthProvider.dart
