+++ auth_provider/lib/src/AuthProviderUser.dart
