+++ auth_provider/lib/src/GraphEmailLoginMethod.dart
