+++ auth_provider/lib/src/GraphEmailSignupMethod.dart
